Renamed file api.js to API.js
